Move jobs into src/
