+++ src/jobs/stalebot.test.ts
@@ -4,12 +4,12 @@ import moment from 'moment-timezone';
 import { GitHubOrg } from '@/api/github/org';
 import { GETSENTRY_ORG, STALE_LABEL } from '@/config';
 
-import { MockedGithubOrg } from '../../../test/utils/testTypes';
+import { MockedGithubOrg } from '../../test/utils/testTypes';
 import {
   WAITING_FOR_COMMUNITY_LABEL,
   WAITING_FOR_PRODUCT_OWNER_LABEL,
   WORK_IN_PROGRESS_LABEL,
-} from '../../config';
+} from '../config';
 
 import { triggerStaleBot } from './stalebot';
